requirements.txt
numpy>=1.20
scipy>=1.6
Pillow>=9.0
matplotlib>=3.5